+++ app/src/main/java/com/sp/smshelper/pdu_utils/AcknowledgeInd.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 /**
  * M-Acknowledge.ind PDU.
