+++ app/src/main/java/com/sp/smshelper/pdu_utils/ContentType.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 import java.util.ArrayList;
