+++ app/src/main/java/com/sp/smshelper/pdu_utils/InvalidHeaderValueException.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 /**
  * Thrown when an invalid header value was set.
