+++ app/src/main/java/com/sp/smshelper/pdu_utils/MmsConfig.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 import android.content.Context;
 import android.content.res.XmlResourceParser;
