+++ app/src/main/java/com/sp/smshelper/pdu_utils/MmsException.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 /**
  * A generic exception that is thrown by the Mms client.
