+++ app/src/main/java/com/sp/smshelper/pdu_utils/MultimediaMessagePdu.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 /**
  * Multimedia message PDU.
