+++ app/src/main/java/com/sp/smshelper/pdu_utils/NotificationInd.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 /**
  * M-Notification.ind PDU.
