+++ app/src/main/java/com/sp/smshelper/pdu_utils/PduCacheEntry.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 public final class PduCacheEntry {
     private final GenericPdu mPdu;
