+++ app/src/main/java/com/sp/smshelper/pdu_utils/PduHeaders.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 import java.util.ArrayList;
 import java.util.HashMap;
