+++ app/src/main/java/com/sp/smshelper/pdu_utils/QuotedPrintable.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 import java.io.ByteArrayOutputStream;
