+++ app/src/main/java/com/sp/smshelper/pdu_utils/SendConf.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 public class SendConf extends GenericPdu {
     /**
