+++ app/src/main/java/com/sp/smshelper/pdu_utils/Utils.java
@@ -1,4 +1,4 @@
-package com.sp.smshelper.utils;
+package com.sp.smshelper.pdu_utils;
 
 import android.Manifest;
 import android.content.Context;
